Enhance physics and movement engine with vehicle control and position tracking
  Major improvements to vehicle physics, entity tracking, and movement execution:

  Core Physics & Movement:
  - Refactor physics_executor for improved vehicle movement simulation
  - Enhanced movement packet sender with better state management
  - Improved physics constants and state tracking for vehicles
  - Add entity callback registry for better event handling

  Entity & Position Tracking:
  - Add entity_position_tracker utility for tracking entity state changes
  - Add ring_buffer utility for efficient position history
  - Improve position getter for mounted entities
  - Enhanced entity callback handling in agent handlers

  Vehicle Enhancements:
  - Implement vehicle physics across all protocol versions (1.21.1-1.21.11)
  - Add minecart steering test suite and common vehicle test utilities
  - Improve boat and vehicle mounting logic
  - Enhanced velocity and physics state management

  Testing & Infrastructure:
  - Remove legacy mock_handlers_test.go
  - Expand vehicle test coverage with common test utilities
  - Add minecart steering validation tests
  - Improve integration test framework for vehicle scenarios

  Code Quality:
  - Update test helpers and agent adapters
  - Clean up replay mirror implementation
  - Improve tracking and state management
  - Enhance command and action handling

+++ agent/actions.go
@@ -389,7 +389,7 @@ func (a *agent) TurnTowards(ctx context.Context, x, y, z float64) error {
 
 	// Calculate yaw using physics formula: atan2(dZ, dX) - 90
 	// This matches YawForStartTarget and ensures consistency
-	yaw := float32(physics.YawForStartTarget(botOrigin, targetPos))
+	yaw := physics.YawForStartTarget(botOrigin, targetPos)
 
 	// Calculate pitch based on arrow spawn height (1.52, not 1.62)
 	// Arrow spawns at: eye - 0.1 = (standing height 1.62) - 0.1 = 1.52
@@ -397,7 +397,7 @@ func (a *agent) TurnTowards(ctx context.Context, x, y, z float64) error {
 	dx := x - botX
 	dz := z - botZ
 	horizontalDist := math.Sqrt(dx*dx + dz*dz)
-	pitch := float32(-math.Atan2(dy, horizontalDist) * 180 / math.Pi)
+	pitch := -math.Atan2(dy, horizontalDist) * 180 / math.Pi
 
 	// Send position and rotation to update body orientation
 	return a.moveExec.SendPositionAndRotation(botX, botY, botZ, yaw, pitch, true)
